pyproject.toml
[tool.poetry]
name = "proximium"
version = "0.1.0"
description = ""
authors = ["Goutham Krishna <gauthamkrishna9991@live.com>"]
readme = "README.md"

[tool.poetry.scripts]
proximium = "proximium.main:main"

[tool.poetry.dependencies]
python = "^3.8"
cassandra-driver = "^3.25.0"
bcrypt = "^3.2.0"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
jwcrypto = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
